Make sure socket action has session even when handling disconnect
When handling disconnect, the handshake object is not available.
We save the last session into the socket object on every request.
This way we can still get to it by loading from the socket object.

This will probably have to be cleaned up: move all data from the
handshake into the socket itself on the first request. Compared
to that, current solution is a hack.